requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.10.0
pydantic-ai>=0.0.12
python-multipart>=0.0.6
reportlab>=4.0.7
openai>=1.3.0
python-dotenv>=1.0.0